Unifico componente de filtros
    refs:#66
