+++ src/components/searchpage/filters/SeriesFilters.tsx
@@ -1,23 +1,37 @@
 import * as React from 'react';
-import { ISerieApi } from '../../../api/SerieApi';
+import { connect } from "react-redux";
+import { IAggregationValue } from '../../../api/ITSAPIResponse';
+import { IStore } from "../../../store/initialState";
+import SearchFilter from '../../common/selector/SearchFilter';
 import Filters from '../../style/Filters/Filters';
 import FilterSubTitle from '../../style/Filters/FilterSubTitle';
 import FilterTitle from '../../style/Filters/FilterTitle';
-import FilterSources from './FilterSources';
-import FilterThemes from './FilterThemes';
 
 
 interface IFilterProps {
-    seriesApi: ISerieApi;
     onSourcePicked: (source: string) => void;
     onThemePicked: (theme: string) => void;
+    picked: string;
+    sources: IAggregationValue[];
+    themes: IAggregationValue[];
 }
 
-export default (props: IFilterProps) =>
+const SeriesFilters = (props: IFilterProps) =>
     <Filters>
         <FilterTitle>Filtros:</FilterTitle>
         <FilterSubTitle>Fuentes:</FilterSubTitle>
-        <FilterSources seriesApi={props.seriesApi} onSourcePicked={props.onSourcePicked} />
+        <SearchFilter onChange={props.onThemePicked} items={props.themes} selected={props.picked} />
         <FilterSubTitle>Temas:</FilterSubTitle>
-        <FilterThemes seriesApi={props.seriesApi} onThemePicked={props.onThemePicked} />
+        <SearchFilter onChange={props.onSourcePicked} items={props.sources} selected={props.picked} />
     </Filters>
+
+
+function mapStateToProps(state: IStore) {
+    return {
+        picked: state.searchParams.datasetTheme,
+        sources: state.aggregations.dataset_source,
+        themes: state.aggregations.dataset_theme,
+    };
+}
+
+export default connect(mapStateToProps)(SeriesFilters);
